+++ src/core/validators/rules/index.ts
@@ -1,0 +1,12 @@
+export { issue } from "./shared.js";
+export {
+  providerIssues,
+  validateId,
+  validateStatus,
+  validateTextPresence,
+  validateUnchanged
+} from "./basic.js";
+export { validatePlaceholders } from "./placeholders.js";
+export { validateNumbers, validateTechnicalTokens, validateVariables } from "./tokens.js";
+export { validateConstraints } from "./constraints.js";
+export { validateGlossary } from "./glossary.js";